lsxtool/requirements.txt
typer[all]>=0.9.0
rich>=13.0.0
pyyaml>=6.0
requests>=2.28.0
